+++ index.html
@@ -320,6 +320,6 @@
       </div>
     </template>
 
-    <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
